Create new superclass UASDAO
This class will act as the superclass for Integrations and Grants (WIP) to share similar methods, authentication methods and URL building
